Backend code/x/requirements.txt
aiofiles==24.1.0
asyncpg==0.30.0
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
greenlet==3.1.1
html5tagger==1.3.0
httptools==0.6.4
itsdangerous==2.2.0
Jinja2==3.1.4
login==0.0.6
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
sanic==24.6.0
Sanic-Cors==2.2.0
sanic-ext==23.12.0
sanic-routing==23.12.0
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
tracerite==1.1.1
typing_extensions==4.12.2
tzdata==2024.2
websockets==14.1
Werkzeug==3.1.3
